Create rectangle_area
Find the area of a rectangle. Write a function in C named 'find_area' that takes int height and int width as arguments, and returns the int area of the rectangle. Call the function in main, to print the area of a rectangle. My answer is hidden below. Remember you have to at least declare a function before calling it in main.