Create FizzBuzz
Time for some conditionals and (hopefully!) a loop! Your favorite problem, FIZZBUZZ!

Write a program in main (or get fancy and write its own function that you call in main) that prints the ints from 1 to 100. But for multiples of three, print "Fizz" instead of the number, and for the multiples of five, print "Buzz." For numbers that are multiples of both three and five, print "FizzBuzz." (Hint: % and &&).

Your program should output each result on a new line. The expected output for the first few numbers is as follows:


1
2
Fizz
4
Buzz
Fizz
7
8
Fizz
Buzz
11
Fizz
13
14
FizzBuzz
16
...
Implement the FizzBuzz logic to accurately replace multiples of three with "Fizz," multiples of five with "Buzz," and multiples of both three and five with "FizzBuzz."